requirements.txt
requests
chromadb
python-dotenv
pyTelegramBotAPI
